meiduo_mall/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Django==4.2.25
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-haystack==3.0
django-js-asset==2.2.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=336dc47db1955fe7451634a6aacf203887bd98d243e29ceeff7b54213fcf4e5f
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-extensions==0.7.1
drf-haystack==1.9.1
elasticsearch==7.10.1
idna==3.11
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==2.1.5
mutagen==1.47.0
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.52
py3Fdfs==2.2.0
pycparser==2.23
pycryptodomex==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-alipay-sdk==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
QQLoginTool==0.3.0
redis==5.1.1
requests==2.32.3
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.19
uWSGI==2.0.31
vine==5.1.0
wcwidth==0.2.14
django-extensions==3.2.3

